requirements.txt
requests==2.31.0
openai==0.27.8
python-dotenv==1.0.0
Pillow==10.0.0
gradio==3.39.0
replicate==0.11.0